+++ src/hook.ts
@@ -1,14 +1,16 @@
 /*
  * @Author: qiansc
  * @Date: 2019-04-23 11:17:36
- * @Last Modified by: qiansc
- * @Last Modified time: 2019-06-05 10:47:39
+ * @Last Modified by: liangjiaying@baidu.com
+ * @Last Modified time: 2019-08-20 15:53:34
  */
+
 import { File, PluginError } from 'gulp-util';
 import path = require('path');
-import stream = require('readable-stream');
 import { include } from './filter';
+import { parseAbsolute, parseBase } from './moduleID';
 import { Parser } from './parser';
+import stream = require('readable-stream');
 const Transform = stream.Transform;
 
 export function amdWrap(option: IAmdWrap) {
@@ -26,6 +28,8 @@ export function amdWrap(option: IAmdWrap) {
       // 传入baseUrl则moduleid基于baseUrl计算
       const baseUrl = option.baseUrl || file.base;
       const prefix = option.prefix || '';
+      const useMd5 = option.useMd5 || false;
+      // let location = parseBase(file.path);
       if (include(file.path, option.exclude, option.baseUrl)) {
         // 在exlude名单中 do nothing
         // console.log('ignore', file.path);
@@ -34,6 +38,7 @@ export function amdWrap(option: IAmdWrap) {
         const parser = new Parser(file.contents, file.path, baseUrl, prefix, option.moduleId, option.staticBaseUrl);
         parser.hook({
           removeModuleId: include(file.path, option.anonymousModule, option.baseUrl),
+          useMd5,
         });
         file.contents = Buffer.from(parser.getContent());
         file.moduleId = parser.getModuleId();
@@ -62,4 +67,13 @@ interface IAmdWrap {
   confAboutFile?: any;
   /** 静态资源的根目录 */
   staticBaseUrl?: string;
+  /** 生成的ModuleId 是否需要md5后缀来避免其他模块引用 如 @molecule/toptip2_134dfas */
+  useMd5?: any;
+}
+
+interface IAmdWrapCustomOption {
+  /** 自定义moduleID */
+  moduleId: string;
+  /** 自定义module path */
+  path: string;
 }