+++ src/index.ts
@@ -1,9 +1,10 @@
 /*
  * @Author: qiansc
  * @Date: 2019-04-23 11:18:08
- * @Last Modified by: qiansc
- * @Last Modified time: 2019-06-05 17:33:33
+ * @Last Modified by: liangjiaying@baidu.com
+ * @Last Modified time: 2019-08-08 19:20:06
  */
 
 export { amdWrap } from './hook';
 export { absolutize } from './absolutize';
+export { addRequireConfig } from './addRequireConfig';