+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.10](https://github.com/searchfe/gulp-amd-wrap/compare/v1.0.9...v1.0.10) (2019-09-19)
+
+
+### Bug Fixes
+
+* support require dynamic module ([e5836ac](https://github.com/searchfe/gulp-amd-wrap/commit/e5836ac))
+* support require dynamic module ([ac4eb0c](https://github.com/searchfe/gulp-amd-wrap/commit/ac4eb0c))
+
 ## [1.0.9](https://github.com/searchfe/gulp-amd-wrap/compare/v1.0.8...v1.0.9) (2019-09-19)
 
 
@@ -35,7 +43,6 @@
 * trigger patch ([444395e](https://github.com/searchfe/gulp-amd-wrap/commit/444395e))
 
 ## [1.0.4](https://github.com/searchfe/gulp-amd-wrap/compare/v1.0.3...v1.0.4) (2019-06-05)
-
 
 ### Bug Fixes